dockerfiles and some refactoring

+++ scripts/sql/create.sql
@@ -2,6 +2,6 @@ create table employee (
 	id BIGSERIAL NOT NULL,
 	title VARCHAR(100),
 	description TEXT,
-	price FLOAT,
+	price INT,
 	stock BOOL
 );
